Add files via upload
In this project I utilized the supervised learning algorithm from sklearn which is Naive Bayes.
The pkl files are given. The email_process file is also given, but I modified it due to the fact that it had many errors.
The one I wrote was the NB ML mini project Jupyter notebook.